refactor(pluggable-widgets-tools): create a helper file
