+++ configs/jsactions/rollup.config.js
@@ -11,6 +11,7 @@ import { promisify } from "util";
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 import typescript from "@rollup/plugin-typescript";
 import { collectDependencies } from "../../packages/tools/pluggable-widgets-tools/configs/rollup-plugin-collect-dependencies";
+import { licenseCustomTemplate } from "../../packages/tools/pluggable-widgets-tools/configs/helpers/rollup-helper";
 import { bigJsImportReplacer } from "./rollup-plugin-bigjs-import-replacer";
 
 const cwd = process.cwd();
@@ -62,18 +63,7 @@ export default async args => {
                                 },
                                 {
                                     file: join(outDir, "dependencies.json"),
-                                    template: dependencies =>
-                                        JSON.stringify(
-                                            dependencies.map(dependency => ({
-                                                [dependency.name]: {
-                                                    version: dependency.version,
-                                                    ...(typeof dependency.isTransitive !== "undefined"
-                                                        ? { transitive: dependency.isTransitive }
-                                                        : null),
-                                                    url: dependency.homepage
-                                                }
-                                            }))
-                                        )
+                                    template: licenseCustomTemplate
                                 }
                             ]
                         }
