+++ packages/tools/pluggable-widgets-tools/configs/rollup.config.js
@@ -30,11 +30,9 @@ import {
     widgetEntry,
     widgetName,
     widgetPackage,
-    widgetVersion,
-    copyLicenseFile,
-    createMpkFile,
-    licenseCustomTemplate
+    widgetVersion
 } from "./shared";
+import { copyLicenseFile, createMpkFile, licenseCustomTemplate } from "./helpers/rollup-helper";
 import url from "./rollup-plugin-assets";
 
 const outDir = join(sourcePath, "/dist/tmp/widgets/");
