+++ packages/tools/pluggable-widgets-tools/configs/shared.js
@@ -1,11 +1,8 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 
-import { existsSync, readdirSync, mkdirSync, promises as fs } from "fs";
+import { existsSync, readdirSync, promises as fs } from "fs";
 import { join } from "path";
 import { config } from "dotenv";
-import fg from "fast-glob";
-import { cp } from "shelljs";
-import { zip } from "zip-a-folder";
 
 config({ path: join(process.cwd(), ".env") });
 
@@ -16,40 +13,6 @@ export async function listDir(path) {
         .map(e => join(path, e.name))
         .concat(...(await Promise.all(entries.filter(e => e.isDirectory()).map(e => listDir(join(path, e.name))))));
 }
-
-export async function copyLicenseFile(sourcePath, outDir) {
-    const absolutePath = join(sourcePath, "licen[cs]e");
-    const licenseFile = (await fg([absolutePath], { cwd: sourcePath, caseSensitiveMatch: false }))[0];
-    if (existsSync(licenseFile)) {
-        cp(licenseFile, outDir);
-    }
-}
-
-export async function createMpkFile(mpkDir, outDir, mpkFile, production, projectPath, deploymentPath) {
-    mkdirSync(mpkDir, { recursive: true });
-    await zip(outDir, mpkFile);
-    if (!production && projectPath) {
-        const widgetsPath = join(projectPath, "widgets");
-        const absolutePath = join(projectPath, deploymentPath);
-        // Create folder if they do not exists or directories were cleaned
-        mkdirSync(widgetsPath, { recursive: true });
-        mkdirSync(absolutePath, { recursive: true });
-        // Copy files to deployment and widgets folder
-        cp("-r", join(outDir, "*"), absolutePath);
-        cp(mpkFile, widgetsPath);
-    }
-}
-
-export const licenseCustomTemplate = dependencies =>
-    JSON.stringify(
-        dependencies.map(dependency => ({
-            [dependency.name]: {
-                version: dependency.version,
-                ...(typeof dependency.isTransitive !== "undefined" ? { transitive: dependency.isTransitive } : null),
-                url: dependency.homepage
-            }
-        }))
-    );
 
 export const sourcePath = process.cwd();
